Switch to a command-oriented argument parsing
The options 'help' and 'version' are now recognized as commands. A
default command 'generate' was added, which is used when no other
command is specified. See the update usage help and README.md for more
information.

This patch makes it easier to implement additional commands which have
different required arguments.